refactor(system): share platform-independent service config
