+++ src/system/services.rs
@@ -2,52 +2,13 @@ use std::path::{Path, PathBuf};
 use std::process::{Command, Output};
 
 use eyre::{Result, bail, eyre};
-use indexmap::{IndexMap, IndexSet};
+use indexmap::IndexMap;
 use serde::{Deserialize, Serialize};
 
-use crate::config::{Config, ConfigMap};
+use crate::config::Config;
 use crate::system::resources::{ResourceAction, ResourceId, ResourceOrigin, ResourcePlan};
 
-#[derive(Clone, Copy, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum ServiceState {
-    #[default]
-    Running,
-    Stopped,
-}
-
-#[derive(Clone, Copy, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum ServiceChangeAction {
-    Reload,
-    Restart,
-    #[default]
-    ReloadOrRestart,
-    None,
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
-pub(crate) struct ServiceTomlConfig {
-    #[serde(default)]
-    pub state: ServiceState,
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-    #[serde(default)]
-    pub masked: bool,
-    #[serde(default)]
-    pub on_change: ServiceChangeAction,
-}
-
-impl Default for ServiceTomlConfig {
-    fn default() -> Self {
-        Self {
-            state: ServiceState::Running,
-            enabled: true,
-            masked: false,
-            on_change: ServiceChangeAction::default(),
-        }
-    }
-}
+pub(crate) use super::services_common::*;
 
 #[derive(Clone, Debug)]
 pub(crate) struct ServiceRequest {
@@ -61,37 +22,7 @@ pub(crate) struct ServiceRequest {
     inspection: Option<ServiceInspection>,
 }
 
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub(crate) struct ServiceNotifications {
-    sources: IndexMap<String, IndexSet<ResourceId>>,
-}
-
 impl ServiceNotifications {
-    pub(crate) fn notify_file(&mut self, path: &Path, services: &[String]) {
-        self.notify(ResourceId::new("file", path.to_string_lossy()), services);
-    }
-
-    pub(crate) fn notify_directory(&mut self, path: &Path, services: &[String]) {
-        self.notify(
-            ResourceId::new("directory", path.to_string_lossy()),
-            services,
-        );
-    }
-
-    #[cfg(test)]
-    pub(crate) fn contains(&self, service: &str) -> bool {
-        self.sources.contains_key(service)
-    }
-
-    fn notify(&mut self, source: ResourceId, services: &[String]) {
-        for service in services {
-            self.sources
-                .entry(service.clone())
-                .or_default()
-                .insert(source.clone());
-        }
-    }
-
     fn change_for(&self, request: &ServiceRequest) -> ServiceChange {
         let Some(sources) = self.sources.get(&request.name) else {
             return ServiceChange::default();
@@ -140,50 +71,12 @@ struct ServicePlan {
 }
 
 pub(crate) fn prepare_requests_from_config(config: &Config) -> Result<Vec<ServiceRequest>> {
-    let mut composed: IndexMap<String, (ServiceTomlConfig, ResourceOrigin)> = IndexMap::new();
-    for config_files in config.bootstrap_config_maps() {
-        for (name, declaration) in services_from_config_files(config_files) {
-            if let Some(existing) = composed.get(&name) {
-                if existing.0 == declaration.0 {
-                    continue;
-                }
-                bail!(
-                    "conflicting bootstrap service declarations for {name}\n\n  first:\n    {}\n\n  second:\n    {}",
-                    existing.1.conflict_description(),
-                    declaration.1.conflict_description(),
-                );
-            }
-            composed.insert(name, declaration);
-        }
-    }
-    composed
+    compose_declarations(config)?
         .into_iter()
         .map(|(name, (config, origin))| {
             ServiceRequest::from_toml_with_origin(name, config, Some(origin))
         })
         .collect()
-}
-
-fn services_from_config_files(
-    config_files: &ConfigMap,
-) -> IndexMap<String, (ServiceTomlConfig, ResourceOrigin)> {
-    let mut merged = IndexMap::new();
-    for (path, cf) in config_files {
-        if let Some(bootstrap) = cf.bootstrap_config() {
-            let origin = ResourceOrigin {
-                config: path.clone(),
-                config_root: cf.config_root(),
-                environment: crate::config::environments_for_config_path(path),
-                source: None,
-            };
-            for (name, service) in bootstrap.services {
-                merged
-                    .entry(name)
-                    .or_insert_with(|| (service, origin.clone()));
-            }
-        }
-    }
-    merged
 }
 
 pub(crate) fn requests_from_config(config: &Config) -> Result<Vec<ServiceRequest>> {
@@ -476,40 +369,6 @@ fn instantiated_unit_template(unit: &str) -> Option<String> {
     (!instance.starts_with('.')).then(|| format!("{prefix}@{suffix}"))
 }
 
-pub(crate) fn validate_notifications(
-    files: &[super::managed_files::ManagedFileRequest],
-    directories: &[super::managed_files::ManagedDirectoryRequest],
-    services: &[ServiceRequest],
-) -> Result<()> {
-    let configured = services
-        .iter()
-        .map(|service| service.name.as_str())
-        .collect::<IndexSet<_>>();
-    for (resource, notification) in files
-        .iter()
-        .flat_map(|file| {
-            file.notify
-                .iter()
-                .map(move |notification| (file.path.as_path(), notification))
-        })
-        .chain(directories.iter().flat_map(|directory| {
-            directory
-                .notify
-                .iter()
-                .map(move |notification| (directory.path.as_path(), notification))
-        }))
-    {
-        if !configured.contains(notification.as_str()) {
-            bail!(
-                "managed path '{}' notifies unconfigured bootstrap service '{}'",
-                resource.display(),
-                notification
-            );
-        }
-    }
-    Ok(())
-}
-
 pub(crate) fn plans_with_notifications(
     requests: &[ServiceRequest],
     notifications: &ServiceNotifications,
@@ -799,10 +658,6 @@ fn unit_file_state_is_enableable(state: &str) -> bool {
         state,
         "disabled" | "enabled" | "enabled-runtime" | "linked" | "linked-runtime" | "indirect"
     )
-}
-
-fn default_true() -> bool {
-    true
 }
 
 #[cfg(test)]
